requirements.txt
pyyaml~=5.3.1
seaborn~=0.10.1
matplotlib~=3.3.0
gym~=0.17.2
numpy~=1.19.1
scipy~=1.5.2
torchvision~=0.5.0
pillow~=7.2.0
ray[rllib]~=1.0.1
opencv-python~=4.4.0.46
future~=0.18.2
torch~=1.4.0
